+++ packages/blob/src/helpers.ts
@@ -1,4 +1,11 @@
-import { type BlobCommandOptions } from '.';
+// common util interface for blob raw commands, not meant to be used directly
+
+import type { PutBlobResult } from '.';
+
+// this is why it's not exported from index/client-upload
+export interface BlobCommandOptions {
+  token?: string;
+}
 
 export function getTokenFromOptionsOrEnv(options?: BlobCommandOptions): string {
   if (options?.token) {
@@ -33,3 +40,41 @@ export class BlobUnknownError extends Error {
     super('Vercel Blob: Unknown error, please visit https://vercel.com/help');
   }
 }
+
+// This version is used to ensure that the client and server are compatible
+// The server (Vercel Blob API) uses this information to change its behavior like the
+// response format
+const BLOB_API_VERSION = 2;
+
+export function getApiVersionHeader(): { 'x-api-version'?: string } {
+  let versionOverride = null;
+  try {
+    // wrapping this code in a try/catch as this function is used in the browser and Vite doesn't define the process.env.
+    // As this varaible is NOT used in production, it will always default to the BLOB_API_VERSION
+    versionOverride =
+      process.env.VERCEL_BLOB_API_VERSION_OVERRIDE ||
+      process.env.NEXT_PUBLIC_VERCEL_BLOB_API_VERSION_OVERRIDE;
+  } catch {
+    // noop
+  }
+
+  return {
+    'x-api-version': `${versionOverride ?? BLOB_API_VERSION}`,
+  };
+}
+
+export function getApiUrl(pathname = ''): string {
+  let baseUrl = null;
+  try {
+    // wrapping this code in a try/catch as this function is used in the browser and Vite doesn't define the process.env.
+    // As this varaible is NOT used in production, it will always default to production endpoint
+    baseUrl =
+      process.env.VERCEL_BLOB_API_URL ||
+      process.env.NEXT_PUBLIC_VERCEL_BLOB_API_URL;
+  } catch {
+    // noop
+  }
+  return `${baseUrl || 'https://blob.vercel-storage.com'}${pathname}`;
+}
+
+export type PutBlobApiResponse = PutBlobResult;
