+++ test/next/src/app/vercel/blob/api/app/handle-blob-upload/edge/route.ts
@@ -1,5 +1,5 @@
-import { handleClientUploadHandler } from '@/app/vercel/blob/handle-blob-upload';
+import { handleUploadHandler } from '@/app/vercel/blob/handle-blob-upload';
 
-export const POST = handleClientUploadHandler;
+export const POST = handleUploadHandler;
 
 export const runtime = 'edge';
